Implement ATOM model API with training, prediction, and testing endpoints
