+++ model/ATOM-Model/app/utils/logger.py
@@ -1,5 +1,8 @@
 import logging
 
 # Then your configuration
-logging.basicConfig(level=logging.INFO)
+logging.basicConfig(
+    level=logging.INFO,
+    format='%(levelname)s:%(message)s'
+)
 logger = logging.getLogger(__name__)
